fix(core): fair loadpoint scheduler — no starvation under pending work
