+++ core/actuationgate_test.go
@@ -154,29 +154,5 @@ func TestScalePhasesGateBothDirections(t *testing.T) {
 	}
 }
 
-// TestLoopLoadpointsPrefersPending verifies the scheduler hands out a pending
-// loadpoint before advancing the round-robin cursor.
-func TestLoopLoadpointsPrefersPending(t *testing.T) {
-	lp0 := &Loadpoint{log: util.NewLogger("lp0")}
-	lp1 := &Loadpoint{log: util.NewLogger("lp1")}
-
-	site := &Site{
-		log:        util.NewLogger("site"),
-		loadpoints: []*Loadpoint{lp0, lp1},
-	}
-
-	lp1.pendingControl.Store(true)
-
-	next := make(chan updater)
-	go site.loopLoadpoints(next)
-
-	if got := <-next; got != lp1 {
-		t.Fatalf("first dispatch: want pending lp1, got %v", got)
-	}
-	if lp1.pendingControl.Load() {
-		t.Fatal("pending flag must be cleared on dispatch")
-	}
-	if got := <-next; got != lp0 {
-		t.Fatalf("second dispatch: want round-robin lp0, got %v", got)
-	}
-}
+// loopLoadpoints scheduler tests live in scheduler_test.go (the scheduler
+// outlives the actuation gate, which is removed in a later phase).
